requirements.txt
Flask==2.3.3
google-cloud-speech==2.27.0
transformers==4.41.0
scikit-learn==1.3.0
nltk==3.8.1
torch==2.5.0
numpy==1.24.3
python-dotenv==1.0.1
tqdm==4.66.5
joblib==1.4.2
huggingface-hub==0.26.0
pymongo==4.5.0
